LayoutTests/fast/scrolling/mac/adjust-scroll-snap-after-gesture-expected.txt



Scrolling
PASS scroller.scrollLeft is 258
Scroller should re-snap
PASS scroller.scrollLeft is 278
PASS successfullyParsed is true

TEST COMPLETE
